# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /users:
    x-swagger-router-controller: users
    post:
      description: Creates new account
      operationId: createAccount
      parameters:
        - name: user
          in: body
          description: user to add to the database
          required: true
          schema:
            $ref: '#/definitions/CreateUser'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CreateUserSuccess"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /favorites:
    x-swagger-router-controller: favorites
    get:
      description: Returns a list favorites of a client
      operationId: returnFavorites
      parameters:
        - name: user_id
          in: query
          description: the favorite beers of the client
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FavoritesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds a new beer to a clients list of favorites
      operationId: addFavorite
      parameters:
        - name: favorite
          in: body
          description: information on the favorite beer of the client
          required: true
          schema:
            $ref: "#/definitions/AddFavoriteRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddFavoriteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a beer from list of favorites
      operationId: deleteFromFavorites
      parameters:
        - name: id
          description: beer id
          type: string
          in: query
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteFavoriteBeer"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /favorite:
    x-swagger-router-controller: favorite
    get:
      description: returns more detailed information about a specific favorite beer
      operationId: specificFavorite
      parameters:
        - name: name
          in: query
          description: detailed information on this specific beer
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SpecificFavoriteBeer"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /wishlist:
    x-swagger-router-controller: wishlist
    get:
      description: Returns all elements on a client's wishlist
      operationId: returnWishlist
      parameters:
        - name: user_id
          in: query
          description: the wishlist beers of the client
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WishlistResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds a new beer to a client's wishlist
      operationId: addToWishlist
      parameters:
        - name: beer_id
          in: query
          description: the beer client wants to add to wishlist
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AddToWishlistResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a beer from client's wishlist
      operationId: deleteFromWishlist
      parameters:
        - name: beer_id
          description: beer id
          type: string
          in: query
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteWishlistBeer"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /search:
    x-swagger-router-controller: search
    get:
      description: Returns a specific beer from brewerydb
      operationId: searchBeer
      parameters:
      - name: name
        in: query
        description: The name of the beer that was searched
        required: true
        type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SearchBeer"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # /token:
  #   get:
  #     description: gets token to authenticate user
  #     operationId: tokenCheck
  #     parameters:
  #       - name: cookie
  #         in: header
  #         description: gets token from req.swagger.user.cookie to authenticate user
  #         required: false
  #         schema:
  #           $ref: '#/definitions/Cookie'
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           a pointer to a definition
  #           $ref: "#/definitions/Authenticated"
  #       responses may fall through to errors
  #       default:
  #         description: Error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  CreateUser:
    type: object
    required:
      - first_name
      - last_name
      - username
      - email
      - password
    properties:
      first_name:
        type: string
        description: first name of new user
      last_name:
        type: string
        description: last name of new user
      username:
        type: string
        description: username
      email:
        type: string
        description: email
      password:
        type: string
        description: password
  CreateUserSuccess:
    type: object
    required:
      - token
      - first_name
      - last_name
      - username
      - email
    properties:
      token:
        type: string
        description: jwt token
      first_name:
        type: string
        description: first name of new user
      last_name:
        type: string
        description: last name of new user
      username:
        type: string
        description: username
      email:
        type: string
        description: email
  FavoritesResponse:
    required:
      - beers
    properties:
      beers:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            label_url:
              type: string
  AddFavoriteRequest:
    type: object
    required:
      - rating
      - comment
      - user_id
      - beer_id
    properties:
      rating:
        type: string
        description: The users chosen rating of the beer(1-5)
      comment:
        type: string
        description: The users comment about the beer
      user_id:
        type: string
        description: user id of client adding new favorite
      beer_id:
        type: string
        description: brewerydb id of new favorite
  AddFavoriteResponse:
    type: object
    required:
      - beer_name
      - label_url
    properties:
      beer_name:
        type: string
      label_url:
        type: string
  DeleteFavoriteBeer:
    type: object
    required:
      - beer_name
    properties:
      beer_name:
        type: string
  SpecificFavoriteBeer:
    type: object
    required:
      - id
      - beer_name
      - abv
      - ibu
      - label_url
      - description
      - type_name
    properties:
      id:
        type: string
      beer_name:
        type: string
      abv:
        type: string
      ibu:
        type: string
      label_url:
        type: string
      description:
        type: string
      type_name:
        type: string
  WishlistResponse:
    required:
      - beers
    properties:
      beers:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            label_url:
              type: string
  AddToWishlistResponse:
    type: object
    required:
      - id
      - beer_name
      - abv
      - ibu
      - label_url
      - description
      - type_name
    properties:
      id:
        type: string
      beer_name:
        type: string
      abv:
        type: string
      ibu:
        type: string
      label_url:
        type: string
      description:
        type: string
      type_name:
        type: string
  DeleteWishlistBeer:
    type: object
    required:
      - beer_id
    properties:
      beer_id:
        type: string
  SearchBeer:
    type: object
    required:
      - id
      - name
      - abv
      - ibu
      - label_url
      - description
      - type_name
    properties:
      id:
        type: string
      name:
        type: string
      abv:
        type: string
      ibu:
        type: string
      label_url:
        type: string
      description:
        type: string
      type_name:
        type: string
  # Cookie:
  #   type: header
  #   properties:
  #     token:
  #       type: int64
  #       description: cookie token
  # Authenticated:
  #   properties:
  #     message:
  #       type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string